# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Portfolio",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/login": {
            "post": {
                "tags": [
                    "Authorization"
                ],
                "summary": "Post a new data",
                "description": "Post new organization",
                "operationId": "3457cc9d5891cad2e5e7deeff6ab7d96",
                "requestBody": {
                    "description": "Login",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "email",
                                        "format": "text",
                                        "example": "a@a.com"
                                    },
                                    "password": {
                                        "type": "password",
                                        "format": "text",
                                        "example": "010203"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Authorization"
                ],
                "summary": "Post a new data",
                "description": "Post new organization",
                "operationId": "c10b3bf89fdc9e529c64542ba162ec94",
                "requestBody": {
                    "description": "Register",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "text",
                                        "format": "text",
                                        "example": "admin"
                                    },
                                    "email": {
                                        "type": "email",
                                        "format": "text",
                                        "example": "a@a.com"
                                    },
                                    "password": {
                                        "type": "password",
                                        "format": "text",
                                        "example": "010203"
                                    },
                                    "password_confirmation": {
                                        "type": "password",
                                        "format": "text",
                                        "example": "010203"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authorization"
                ],
                "summary": "Post a new data",
                "description": "Logout",
                "operationId": "43f817e8d298bafd1db3953ec98f4cb3",
                "responses": {
                    "422": {
                        "description": "Wrong credentials response"
                    },
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            }
        },
        "/api/refresh": {
            "post": {
                "tags": [
                    "Authorization"
                ],
                "summary": "Post a new data",
                "description": "Reshresh",
                "operationId": "b2e12501184f2fdaca855a0ec45361eb",
                "responses": {
                    "422": {
                        "description": "Wrong credentials response"
                    },
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "api": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}